[mod] query_test.go use case of gin.Context.ShouldBind()

+++ cmd/cache.go
@@ -7,7 +7,7 @@ type (
 	// Key : Key for CacheMap
 	Key struct {
 		Word  string
-		Limit int
+		Limit uint64
 	}
 )
